separate custom admin site into project directory, keep task model and preferences in migrator app
