+++ src/pymap/urls.py
@@ -19,7 +19,7 @@ from django.contrib.auth import views as auth_views
 from django.contrib.auth.decorators import login_required
 from django.conf import settings
 
-from migrator.admin import custom_admin_site
+from .admin import custom_admin_site
 
 urlpatterns = [
     path("admin/doc/", include("django.contrib.admindocs.urls")),
